+++ .build.ps1
@@ -4,40 +4,52 @@
 #>
 
 param(
-	$Configuration = (property Configuration Release)
+	$Configuration = (property Configuration Release),
+	$FarHome = (property FarHome C:\Bin\Far\x64)
 )
 
-Set-StrictMode -Version 2
+Set-StrictMode -Version 3
 $ModuleName = 'FarNet.FSharp.PowerShell'
-$env:FarDevHome = $FarDevHome = if (Test-Path 'C:\Bin\Far\x64') {'C:\Bin\Far\x64'} else {''}
+$ModuleRoot = "$FarHome\FarNet\Lib\$ModuleName"
+$Description = 'F# friendly PowerShell Core helper'
 
-# Synopsis: Remove temp files.
+task build meta, {
+	exec { dotnet build -c $Configuration }
+}
+
+task publish {
+	Set-Location src
+	Copy-Item -Destination $ModuleRoot @(
+		"$ModuleName.ini"
+		"bin\$Configuration\net6.0\$ModuleName.dll"
+		"bin\$Configuration\net6.0\$ModuleName.xml"
+	)
+}
+
 task clean {
 	remove *\bin, *\obj, README.htm, *.nupkg, z
 }
 
-# Synopsis: Build and Post (post build target).
-task build {
-	exec {dotnet build -c $Configuration}
-}
-
-# Synopsis: Post build target. Copy stuff.
-task post -If:$FarDevHome {
-	$to = "$FarDevHome\FarNet\Lib\$ModuleName"
-	Copy-Item "src\$ModuleName.ini" $to
-}
-
-# Get version from release notes.
-function Get-Version {
-	switch -Regex -File Release-Notes.md {'##\s+v(\d+\.\d+\.\d+)' {return $Matches[1]} }
-}
-
-# Synopsis: Set $script:Version.
 task version {
-	($script:Version = Get-Version)
+	($script:Version = switch -Regex -File Release-Notes.md {'##\s+v(\d+\.\d+\.\d+)' {$Matches[1]; break}})
 }
 
-# Synopsis: Convert markdown to HTML.
+task meta -Inputs .build.ps1, Release-Notes.md -Outputs src/Directory.Build.props -Jobs version, {
+	Set-Content src/Directory.Build.props @"
+<Project>
+	<PropertyGroup>
+		<Company>https://github.com/nightroman/$ModuleName</Company>
+		<Copyright>Copyright (c) Roman Kuzmin</Copyright>
+		<Description>$Description</Description>
+		<Product>$ModuleName</Product>
+		<Version>$Version</Version>
+		<FileVersion>$Version</FileVersion>
+		<AssemblyVersion>$Version</AssemblyVersion>
+	</PropertyGroup>
+</Project>
+"@
+}
+
 task markdown {
 	assert (Test-Path $env:MarkdownCss)
 	exec { pandoc.exe @(
@@ -49,66 +61,46 @@ task markdown {
 	)}
 }
 
-# Synopsis: Generate meta files.
-task meta -Inputs .build.ps1, Release-Notes.md -Outputs src/Directory.Build.props -Jobs version, {
-	Set-Content src/Directory.Build.props @"
-<Project>
-	<PropertyGroup>
-		<Company>https://github.com/nightroman/$ModuleName</Company>
-		<Copyright>Copyright (c) Roman Kuzmin</Copyright>
-		<Description>F# friendly PowerShell extension</Description>
-		<Product>$ModuleName</Product>
-		<Version>$Version</Version>
-		<FileVersion>$Version</FileVersion>
-		<AssemblyVersion>$Version</AssemblyVersion>
-	</PropertyGroup>
-</Project>
-"@
-}
-
 # Synopsis: Collect package files.
-task package -If:$FarDevHome markdown, {
+task package markdown, {
 	remove z
-	$toLib = mkdir "z\lib\net472"
+	$toLib = mkdir "z\lib\net6.0"
 	$toModule = mkdir "z\tools\FarHome\FarNet\Lib\$ModuleName"
-	$fromModule = "$FarDevHome\FarNet\Lib\$ModuleName"
 
 	Copy-Item -Destination $toLib @(
-		"$fromModule\FarNet.FSharp.PowerShell.dll"
-		"$fromModule\FarNet.FSharp.PowerShell.xml"
+		"$ModuleRoot\FarNet.FSharp.PowerShell.dll"
+		"$ModuleRoot\FarNet.FSharp.PowerShell.xml"
 	)
 
 	Copy-Item -Destination $toModule @(
 		'README.htm'
 		'LICENSE'
-		"$fromModule\FarNet.FSharp.PowerShell.dll"
-		"$fromModule\FarNet.FSharp.PowerShell.ini"
-		"$fromModule\FarNet.FSharp.PowerShell.xml"
+		"$ModuleRoot\FarNet.FSharp.PowerShell.dll"
+		"$ModuleRoot\FarNet.FSharp.PowerShell.ini"
+		"$ModuleRoot\FarNet.FSharp.PowerShell.xml"
 	)
 }
 
 # Synopsis: Make NuGet package.
-task nuget -If:$FarDevHome package, version, {
+task nuget package, version, {
 	# test versions
-	$dllPath = "$FarDevHome\FarNet\Lib\$ModuleName\$ModuleName.dll"
+	$dllPath = "$FarHome\FarNet\Lib\$ModuleName\$ModuleName.dll"
 	($dllVersion = (Get-Item $dllPath).VersionInfo.FileVersion.ToString())
 	assert $dllVersion.StartsWith("$Version.") 'Versions mismatch.'
 
-	$text = @'
-F# friendly PowerShell extension
+	$Description = @"
+$Description
 
 ---
 
-The package may be used as usual in F# projects.
-
-It is also configured for FarNet.FSharpFar.
+The package is designed for FarNet.FSharpFar and requires FarNet.PowerShellFar.
 To install FarNet packages, follow these steps:
 
 https://github.com/nightroman/FarNet#readme
 
 ---
-'@
-	# nuspec
+"@
+
 	Set-Content z\Package.nuspec @"
 <?xml version="1.0"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
@@ -119,37 +111,37 @@ https://github.com/nightroman/FarNet#readme
 		<owners>Roman Kuzmin</owners>
 		<projectUrl>https://github.com/nightroman/$ModuleName</projectUrl>
 		<license type="expression">Apache-2.0</license>
-		<requireLicenseAcceptance>false</requireLicenseAcceptance>
-		<summary>$text</summary>
-		<description>$text</description>
-		<releaseNotes>https://github.com/nightroman/FarNet.FSharp.PowerShell/blob/master/Release-Notes.md</releaseNotes>
+		<description>$Description</description>
+		<releaseNotes>https://github.com/nightroman/FarNet.FSharp.PowerShell/blob/main/Release-Notes.md</releaseNotes>
 		<tags>FSharp PowerShell FarManager FarNet FSharpFar</tags>
 		<dependencies>
-			<group targetFramework=".NETFramework4.7.2">
-				<dependency id="Microsoft.PowerShell.3.ReferenceAssemblies" version="1.0.0" />
+			<group targetFramework="net6.0">
+				<dependency id="Microsoft.PowerShell.SDK" version="7.2.6" />
 			</group>
 		</dependencies>
 	</metadata>
 </package>
 "@
-	# pack
+
 	exec { NuGet.exe pack z\Package.nuspec }
 }
 
-# Synopsis: xUnit.
+# Synopsis: xUnit. Idle if the main project is clean.
 task test1 {
-	Push-Location tests
-	exec { dotnet test -c $Configuration }
-	Pop-Location
-	remove tests\bin, tests\obj
+	Set-Location tests
+	exec { dotnet test -c $Configuration -l "console;verbosity=normal" --no-build }
 }
 
-# Synopsis: fsx.
-task test2 -If:$FarDevHome {
+# Synopsis: fsx tests.
+task test2 {
 	Invoke-Build ** tests
 }
 
 # Synopsis: All tests.
 task test test1, test2
 
-task . build, test, clean
+# Synopsis: Update and test.
+task all build, test, clean
+
+# Synopsis: Update.
+task . build, clean