+++ tests/.test.ps1
@@ -2,12 +2,12 @@
 # generate tasks for fsx files in /samples
 Get-Item ../samples/*.fsx | .{process{
 	Add-BuildTask -Name:$_.Name -Data:$_ -Jobs:{
-		exec { fsx.exe $Task.Data }
+		exec { fsx $Task.Data }
 	}
 }}
 
 task missingCommand {
-	($r = fsx.exe test.fsx missingCommand)
+	($r = fsx test.fsx missingCommand)
 	equals $LASTEXITCODE 1
 	equals $r $null
 }