+++ tests/Program.fs
@@ -1,5 +1,0 @@
-﻿open System
-
-[<EntryPoint>]
-let main argv =
-    0